requirements.txt
dbt-core==1.8.2
dbt-postgres==1.8.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9